Fix 6 critical bugs found by compound negative audit
- fused_eou: guard NaN/Inf in prosody ring buffer and sigmoid overflow
- fused_eou: apply context_adj to solo trigger thresholds (was bypassed)
- conformer_stt: clear layer caches on Conmer mode switch (stale MHSA data)
- metal_dispatch: atomic CAS on g_initialized (race condition)
- sonata_istft: size_t cast in batch decode pointer arithmetic (overflow)
- Add 6 new audit bug-fix tests (40/40 EOU, 70/70 total research tests)

+++ src/sonata_istft.c
@@ -251,8 +251,8 @@ int sonata_istft_decode_batch(
     for (int f = 0; f < n_frames; f++) {
         int n = sonata_istft_decode_frame(
             dec,
-            magnitudes + f * dec->n_bins,
-            phases + f * dec->n_bins,
+            magnitudes + (size_t)f * dec->n_bins,
+            phases + (size_t)f * dec->n_bins,
             out_audio + total_samples
         );
         total_samples += n;
